Fixed dumpgen to work and not recursivly crash

+++ src/shared/utilcore/code/CrashDumpThread.cpp
@@ -152,7 +152,7 @@ void CrashDumpThread::uploadProgress(Prog_s& prog)
 	if (m_bSecond)
 		total += 50;
 
-	total += prog.ulnow*50/prog.ultotal;
+	total += (uint8)(prog.ulnow*50.0/prog.ultotal);
 	onProgressEvent(total);
 
 	prog.abort = isStopped();
